Refactor: split up syscalls module
